backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
beautifulsoup4
requests
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
